requirements.txt
absl-py==2.1.0
appdirs==1.4.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.60.1
gymnasium==0.29.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.16.0
tensorboard-data-server==0.7.2
tf_keras-nightly==2.16.0.dev2024021310
threadpoolctl==3.3.0
torch==2.2.0
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.3
Werkzeug==3.0.1
